+++ src/features/library/components/LibraryOptionsPanel.tsx
@@ -11,6 +11,9 @@ import FormLabel from '@mui/material/FormLabel';
 import RadioGroup from '@mui/material/RadioGroup';
 import { useLingui } from '@lingui/react/macro';
 import { msg } from '@lingui/core/macro';
+import type { ReactNode } from 'react';
+import { useMemo } from 'react';
+import uniqBy from 'lodash/fp/uniqBy';
 import { CheckboxInput } from '@/base/components/inputs/CheckboxInput.tsx';
 import { RadioInput } from '@/base/components/inputs/RadioInput.tsx';
 import { SortRadioInput } from '@/base/components/inputs/SortRadioInput.tsx';
@@ -34,6 +37,8 @@ import type { CategoryMetadataInfo } from '@/features/category/Category.types.ts
 import { MANGA_STATUS_TO_TRANSLATION } from '@/features/manga/Manga.constants.ts';
 import { GridLayout } from '@/base/Base.types';
 import { getErrorMessage } from '@/lib/HelperFunctions.ts';
+import { Collapsable } from '@/base/components/Collapsable.tsx';
+import Replay from '@mui/icons-material/Replay';
 
 const TITLES: { [key in 'filter' | 'sort' | 'display']: MessageDescriptor } = {
     filter: msg`Filter`,
@@ -49,27 +54,49 @@ const SORT_OPTIONS: [LibrarySortMode, MessageDescriptor][] = [
     ['lastRead', msg`Recently read`],
     ['latestFetchedChapter', msg`Latest fetched chapter`],
     ['latestUploadedChapter', msg`Latest uploaded chapter`],
+    ['random', msg`Random`],
 ];
+
+const CollapsableFilter = ({ title, items, isActive }: { title: string; items: ReactNode[]; isActive: boolean }) => (
+    <Collapsable
+        header={title}
+        collapse={items}
+        initialState={isActive}
+        slots={{
+            headerWrapper: {
+                component: FormLabel,
+                sx: {
+                    mt: 2,
+                    color: isActive ? 'warning.main' : undefined,
+                },
+            },
+        }}
+    />
+);
 
 export const LibraryOptionsPanel = ({
     category,
     open,
     onClose,
-    uniqueSources = STABLE_EMPTY_ARRAY,
-    sourceFilters = {},
-    onSourceChange,
 }: {
     category: CategoryMetadataInfo;
     open: boolean;
     onClose: () => void;
-    uniqueSources?: Array<{ id: string; name: string; displayName?: string }>;
-    sourceFilters?: Record<string, boolean> | null;
-    onSourceChange: (sourceId: string, checked: boolean | null) => void;
 }) => {
     const { t } = useLingui();
 
     const trackerList = requestManager.useGetTrackerList<GetTrackersSettingsQuery>(GET_TRACKERS_SETTINGS);
     const loggedInTrackers = Trackers.getLoggedIn(trackerList.data?.trackers.nodes ?? STABLE_EMPTY_ARRAY);
+
+    const migratableSourcesResult = requestManager.useGetMigratableSources();
+    const librarySources = useMemo(() => {
+        const sources = migratableSourcesResult.data?.mangas.nodes
+            .map(({ source }) => source)
+            .filter((source) => source != null);
+        const uniqueSources = uniqBy('id', sources);
+
+        return uniqueSources.toSorted((a, b) => a.displayName.localeCompare(b.displayName));
+    }, [migratableSourcesResult.data?.mangas.nodes]);
 
     const categoryLibraryOptions = useGetCategoryMetadata(category);
     const updateCategoryLibraryOptions = createUpdateCategoryMetadata(category, (e) =>
@@ -89,6 +116,14 @@ export const LibraryOptionsPanel = ({
     const setSettingValue = createUpdateMetadataServerSettings((e) =>
         makeToast(t`Could not save the default search settings to the server`, 'error', getErrorMessage(e)),
     );
+
+    const isStatusFilterActive = Object.values(MangaStatus).some(
+        (status) => categoryLibraryOptions.hasStatus[status] != null,
+    );
+    const isTrackerFilterActive = loggedInTrackers.some(
+        (tracker) => categoryLibraryOptions.hasTrackerBinding[tracker.id] != null,
+    );
+    const isSourceFilterActive = librarySources.some((source) => categoryLibraryOptions.hasSource[source.id] != null);
 
     return (
         <OptionsTabs<'filter' | 'sort' | 'display'>
@@ -125,47 +160,61 @@ export const LibraryOptionsPanel = ({
                                 checked={categoryLibraryOptions.hasDuplicateChapters}
                                 onChange={(c) => updateCategoryLibraryOptions('hasDuplicateChapters', c)}
                             />
-                            <FormLabel sx={{ mt: 2 }}>{t`Status`}</FormLabel>
-                            {Object.values(MangaStatus).map((status) => (
-                                <ThreeStateCheckboxInput
-                                    key={status}
-                                    label={t(MANGA_STATUS_TO_TRANSLATION[status])}
-                                    checked={categoryLibraryOptions.hasStatus[status]}
-                                    onChange={(checked) =>
-                                        updateCategoryLibraryOptions('hasStatus', {
-                                            ...categoryLibraryOptions.hasStatus,
-                                            [status]: checked,
-                                        })
-                                    }
-                                />
-                            ))}
-                            <FormLabel sx={{ mt: 2 }}>{t`Tracked`}</FormLabel>
-                            {loggedInTrackers.map((tracker) => (
-                                <ThreeStateCheckboxInput
-                                    key={tracker.id}
-                                    label={tracker.name}
-                                    checked={categoryLibraryOptions.hasTrackerBinding[tracker.id]}
-                                    onChange={(checked) =>
-                                        updateCategoryLibraryOptions('hasTrackerBinding', {
-                                            ...categoryLibraryOptions.hasTrackerBinding,
-                                            [tracker.id]: checked,
-                                        })
-                                    }
-                                />
-                            ))}
-                            <FormLabel sx={{ mt: 2 }}>{t`Sources`}</FormLabel>
-                            {uniqueSources.map((source) => {
-                                const currentStatus = sourceFilters?.[source.id] ?? null;
-
-                                return (
+                            <CollapsableFilter
+                                title={t`Status`}
+                                items={Object.values(MangaStatus).map((status) => (
                                     <ThreeStateCheckboxInput
-                                        key={source.id}
-                                        label={source.displayName || source.name}
-                                        checked={currentStatus}
-                                        onChange={(checked) => onSourceChange(source.id, checked ?? null)}
+                                        key={status}
+                                        label={t(MANGA_STATUS_TO_TRANSLATION[status])}
+                                        checked={categoryLibraryOptions.hasStatus[status]}
+                                        onChange={(checked) =>
+                                            updateCategoryLibraryOptions('hasStatus', {
+                                                ...categoryLibraryOptions.hasStatus,
+                                                [status]: checked,
+                                            })
+                                        }
                                     />
-                                );
-                            })}
+                                ))}
+                                isActive={isStatusFilterActive}
+                            />
+                            {!!loggedInTrackers.length && (
+                                <CollapsableFilter
+                                    title={t`Tracked`}
+                                    items={loggedInTrackers.map((tracker) => (
+                                        <ThreeStateCheckboxInput
+                                            key={tracker.id}
+                                            label={tracker.name}
+                                            checked={categoryLibraryOptions.hasTrackerBinding[tracker.id]}
+                                            onChange={(checked) =>
+                                                updateCategoryLibraryOptions('hasTrackerBinding', {
+                                                    ...categoryLibraryOptions.hasTrackerBinding,
+                                                    [tracker.id]: checked,
+                                                })
+                                            }
+                                        />
+                                    ))}
+                                    isActive={isTrackerFilterActive}
+                                />
+                            )}
+                            {!!librarySources.length && (
+                                <CollapsableFilter
+                                    title={t`Source`}
+                                    items={librarySources.map((source) => (
+                                        <ThreeStateCheckboxInput
+                                            key={source.id}
+                                            label={source.displayName}
+                                            checked={categoryLibraryOptions.hasSource[source.id]}
+                                            onChange={(checked) =>
+                                                updateCategoryLibraryOptions('hasSource', {
+                                                    ...categoryLibraryOptions.hasSource,
+                                                    [source.id]: checked,
+                                                })
+                                            }
+                                        />
+                                    ))}
+                                    isActive={isSourceFilterActive}
+                                />
+                            )}
                         </>
                     );
                 }
@@ -175,6 +224,7 @@ export const LibraryOptionsPanel = ({
                             key={mode}
                             label={t(label)}
                             checked={categoryLibraryOptions.sortBy === mode}
+                            checkedIcon={mode === 'random' ? <Replay color="primary" /> : undefined}
                             sortDescending={categoryLibraryOptions.sortDesc}
                             onClick={() =>
                                 mode !== categoryLibraryOptions.sortBy